+++ common/inc/sfpu/ckernel_sfpu_abs.h
@@ -1,0 +1,33 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ckernel_defs.h"
+#include "ckernel.h"
+#include "noc_nonblocking_api.h"
+
+#include "sfpi.h"
+
+using namespace sfpi;
+
+namespace ckernel
+{
+namespace sfpu
+{
+
+template <bool APPROXIMATION_MODE, int ITERATIONS>
+inline void _calculate_abs_(const int iterations)
+{
+    // SFPU microcode
+    for (int d = 0; d < iterations; d++)
+    {
+        vFloat v = dst_reg[0];
+        dst_reg[0] = sfpi::abs(v);
+        dst_reg++;
+    }
+}
+
+} // namespace sfpu
+} // namespace ckernel
