+++ common/inc/sfpu/ckernel_sfpu_comp.h
@@ -1,0 +1,98 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ckernel_defs.h"
+#include "ckernel.h"
+#include "noc_nonblocking_api.h"
+#include "ckernel_sfpu_is_fp16_zero.h"
+
+#include "sfpi.h"
+
+using namespace sfpi;
+
+namespace ckernel
+{
+namespace sfpu
+{
+
+sfpi_inline void _calculate_comp_init_flag_(bool check, vFloat& flag1, vFloat& flag2, float init)
+{
+    flag1 = init;
+    if (check) {
+        flag2 = init;
+    }
+}
+
+template <bool APPROXIMATION_MODE, bool invert_output, bool check_zero, bool second_check, bool is_less_than_equal_zero, int ITERATIONS>
+inline void _calculate_comp_(const int iterations, uint exponent_size_8)
+{
+
+    // output_0 and output_1 hold the outputs use use when a zero or negative check is true/false.
+    // False = 0.0 = kCONST_0 (5/8-bit exponent format)
+    // True  = 1.0 = kCONST_1_FP16B (8-bit exponent format)
+    // SFPU uses 8-bit exponent in operations so loading these constants in 8-bit exponent format.
+    // Although a command flag can tell SFPU to re-bias a 5-bit exponent to 8-bit, we are loading 8-bit
+    // exponent and telling SFPU to not add any bias to these constants.
+    constexpr float output_0 = invert_output ? 0.0f : 1.0f;
+    constexpr float output_1 = invert_output ? 1.0f : 0.0f;
+
+    for (int d = 0; d < iterations; d++)
+    {
+        vFloat v = dst_reg[0];
+        vFloat flag1, flag2;
+        if constexpr(check_zero)
+        {
+            v_if (_sfpu_is_fp16_zero_(v, exponent_size_8)) {
+                _calculate_comp_init_flag_(second_check, flag1, flag2, output_0);
+            } v_else {
+                _calculate_comp_init_flag_(second_check, flag1, flag2, output_1);
+            }
+            v_endif;
+        }
+        else
+        {
+            v_if (v < 0.0F) {
+                _calculate_comp_init_flag_(second_check, flag1, flag2, output_0);
+            } v_else {
+                _calculate_comp_init_flag_(second_check, flag1, flag2, output_1);
+            }
+            v_endif;
+        }
+
+        vFloat result;
+        if constexpr (second_check)
+        {
+            // less_than_equal_zero
+            // flag1 = 0x3F80(1.0) if DST < 0 else 0
+            // flag2 = 0x3F80(1.0) if DST == 0 else 0
+            // Do a bitwise Or (flag1 | flag2) to get <= condition.
+            // flag1 < 0 OR flag2 == 0 => DST is Less than or Equal to zero.
+            // Result will be either 0x0000(0.0) or 0x3F80(1.0)
+            if constexpr (is_less_than_equal_zero) {
+                result = reinterpret<vFloat>(reinterpret<vUInt>(flag1) | reinterpret<vUInt>(flag2));
+            }
+            else
+            {
+                // greater_than_zero
+                // flag1 = 0x3F80(1.0) if DST >= 0 else 0
+                // flag2 = 0x3F80(1.0) if DST != 0 else 0
+                // Do a bitwise And (flag1 & flag2) to get > condition.
+                // flag2 >= 0 AND flag1 != 0 => DST is Greater than zero
+                // Result will be either 0x0000(0.0) or 0x3F80(1.0)
+                result = reinterpret<vFloat>(reinterpret<vUInt>(flag1) & reinterpret<vUInt>(flag2));
+            }
+        } else {
+            result = flag1;
+        }
+
+        dst_reg[0] = result;
+
+        dst_reg++;
+    }
+}
+
+} // namespace sfpu
+} // namespace ckernel
