+++ common/inc/sfpu/ckernel_sfpu_sign.h
@@ -1,0 +1,48 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ckernel_defs.h"
+#include "ckernel.h"
+#include "noc_nonblocking_api.h"
+#include "ckernel_sfpu_is_fp16_zero.h"
+
+#include "sfpi.h"
+
+using namespace sfpi;
+
+namespace ckernel
+{
+namespace sfpu
+{
+
+template <bool APPROXIMATION_MODE, int ITERATIONS>
+inline void _calculate_sign_(const int iterations, uint exponent_size_8)
+{
+    // All params are in FP16 format
+    // uint format = 1;
+    #pragma GCC unroll 0
+    for (int d = 0; d < iterations; d++)
+    {
+        vFloat v = dst_reg[0];
+        dst_reg[0] = vConst1;
+        v_if (v < 0.0F) {
+            dst_reg[0] = vConstNeg1;
+        }
+        v_endif;
+
+        //param0 == 0 is Bfp8 format. It does not require bias removal.
+        //param0 != 0 is Float16 format and exp bias needs to be removed for zero check.
+        v_if (_sfpu_is_fp16_zero_(v, exponent_size_8)) {
+            dst_reg[0] = vConst0;
+        }
+        v_endif;
+
+        dst_reg++;
+    }
+}
+
+} // namespace sfpu
+} // namespace ckernel
