+++ common/inc/sfpu/ckernel_sfpu_tanh_derivative.h
@@ -1,0 +1,48 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ckernel_defs.h"
+#include "ckernel.h"
+#include "noc_nonblocking_api.h"
+
+#include "sfpi.h"
+
+using namespace sfpi;
+
+namespace ckernel
+{
+namespace sfpu
+{
+
+template <bool APPROXIMATION_MODE, int WITH_PRECOMPUTED_TANH, int ITERATIONS>
+inline void _calculate_tanh_derivative_(const int iterations)
+{
+    vUInt l0 = l_reg[LRegs::LReg0];
+    vUInt l1 = l_reg[LRegs::LReg1];
+    vUInt l2 = l_reg[LRegs::LReg2];
+
+    // tanh'(x) = 1 - (tanh(x))^2
+    for (int d = 0; d < iterations; d++)
+    {
+        vFloat val = dst_reg[0];
+
+        if constexpr (!WITH_PRECOMPUTED_TANH) {
+            val = lut(val, l0, l1, l2);
+        }
+
+        val = val * (-val) + vConst1;
+        dst_reg[0] = val;
+
+        dst_reg++;
+    }
+
+    l_reg[LRegs::LReg0] = l0;
+    l_reg[LRegs::LReg1] = l1;
+    l_reg[LRegs::LReg2] = l2;
+}
+
+} // namespace sfpu
+} // namespace ckernel
